Split lexer string interpolation scanning
