+++ src/lexer/strings.rs
@@ -128,182 +128,4 @@ impl Lexer {
             Ok(first)
         }
     }
-
-    /// Lex tokens inside `${...}`, tracking brace depth.
-    /// Pushes all tokens (including the closing InterpolationEnd) into `buf`.
-    fn lex_interpolation_body(&mut self, buf: &mut Vec<(Token, Span)>) -> Result<(), CompileError> {
-        let mut depth = 1u32;
-
-        loop {
-            // Inside interpolation, newlines are insignificant
-            self.skip_all_whitespace_and_comments()?;
-
-            match self.peek() {
-                None => {
-                    return Err(CompileError::Syntax(
-                        "unterminated string interpolation".into(),
-                        Some(self.make_span(self.byte_pos(), self.byte_pos())),
-                    ));
-                }
-                Some('}') => {
-                    depth -= 1;
-                    if depth == 0 {
-                        let s = self.byte_pos();
-                        self.advance();
-                        buf.push((Token::InterpolationEnd, self.make_span(s, self.byte_pos())));
-                        return Ok(());
-                    }
-                    // Nested } — emit as RBrace
-                    let s = self.byte_pos();
-                    self.advance();
-                    buf.push((Token::RBrace, self.make_span(s, self.byte_pos())));
-                }
-                Some('{') => {
-                    depth += 1;
-                    let s = self.byte_pos();
-                    self.advance();
-                    buf.push((Token::LBrace, self.make_span(s, self.byte_pos())));
-                }
-                Some('"') => {
-                    // Nested string inside interpolation — lex it recursively.
-                    // lex_string may set self.pending, so we save and restore.
-                    let saved_pending = std::mem::take(&mut self.pending);
-                    let first = self.lex_string()?;
-                    buf.push(first);
-                    buf.append(&mut self.pending);
-                    self.pending = saved_pending;
-                }
-                _ => {
-                    // Use lex_next for a regular token (skip_whitespace already done)
-                    let (tok, span) = self.lex_next_no_skip()?;
-                    if tok.is_eof() {
-                        return Err(CompileError::Syntax(
-                            "unterminated string interpolation".into(),
-                            Some(span),
-                        ));
-                    }
-                    buf.push((tok, span));
-                }
-            }
-        }
-    }
-
-    /// Lex a single token without calling skip_whitespace first.
-    /// Used inside interpolation body where we already skipped whitespace.
-    fn lex_next_no_skip(&mut self) -> Result<(Token, Span), CompileError> {
-        let start = self.byte_pos();
-
-        let ch = match self.peek() {
-            Some(c) => c,
-            None => return Ok((Token::EOF, self.make_span(start, start))),
-        };
-
-        // Identifiers / keywords
-        if ch.is_ascii_alphabetic() || ch == '_' {
-            return Ok(self.lex_identifier());
-        }
-
-        // Numbers
-        if ch.is_ascii_digit() {
-            return self.lex_number();
-        }
-
-        // @ tokens
-        if ch == '@' {
-            return self.lex_at_token();
-        }
-
-        // Multi-char operators
-        if self.matches("::=") {
-            self.advance_n(3);
-            return Ok((Token::DeclareAssign, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches(":=") {
-            self.advance_n(2);
-            return Ok((Token::ConstAssign, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("..=") {
-            self.advance_n(3);
-            return Ok((Token::DotDotEq, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("..") {
-            self.advance_n(2);
-            return Ok((Token::DotDot, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("=>") {
-            self.advance_n(2);
-            return Ok((Token::FatArrow, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("->") {
-            self.advance_n(2);
-            return Ok((Token::Arrow, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("==") {
-            self.advance_n(2);
-            return Ok((Token::Eq, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("!=") {
-            self.advance_n(2);
-            return Ok((Token::NotEq, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("<=") {
-            self.advance_n(2);
-            return Ok((Token::LtEq, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches(">=") {
-            self.advance_n(2);
-            return Ok((Token::GtEq, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("&&") {
-            self.advance_n(2);
-            return Ok((Token::And, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("||") {
-            self.advance_n(2);
-            return Ok((Token::Or, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches("<<") {
-            self.advance_n(2);
-            return Ok((Token::ShiftLeft, self.make_span(start, self.byte_pos())));
-        }
-        if self.matches(">>") {
-            self.advance_n(2);
-            return Ok((Token::ShiftRight, self.make_span(start, self.byte_pos())));
-        }
-
-        // Single-char
-        self.advance();
-        let end = self.byte_pos();
-        let tok = match ch {
-            '+' => Token::Plus,
-            '-' => Token::Minus,
-            '*' => Token::Star,
-            '/' => Token::Slash,
-            '%' => Token::Percent,
-            '<' => Token::Lt,
-            '>' => Token::Gt,
-            '!' => Token::Not,
-            '&' => Token::BitAnd,
-            '|' => Token::Pipe,
-            '^' => Token::BitXor,
-            '~' => Token::Tilde,
-            '=' => Token::Assign,
-            '.' => Token::Dot,
-            '?' => Token::Question,
-            '(' => Token::LParen,
-            ')' => Token::RParen,
-            '[' => Token::LBracket,
-            ']' => Token::RBracket,
-            ',' => Token::Comma,
-            ':' => Token::Colon,
-            ';' => Token::Semicolon,
-            _ => {
-                return Err(CompileError::Syntax(
-                    format!("unexpected character '{ch}'"),
-                    Some(self.make_span(start, end)),
-                ));
-            }
-        };
-        Ok((tok, self.make_span(start, end)))
-    }
 }
